Add support for OSX (#84)
No new features. Creates an environment file and down-selects from the testing scripts to only those which run on OS X easily. E.g., there is interference xTB and Tensorflow interfere with other packages, so I don't test them.

* Test on MacOS

* Try using a regular Conda not Mamba

* Install CP2K with brew

* Try Conda for both OSX and Linux installs

* Extend the action timeout

We're also using the timeouts with pytest

* Use mamba on ubuntu

Not sure if I have the syntax correct

* Use single quotes

* Check to make sure the doer launches

* Skip the CLI component on Mac

Not because I won't fix it, but because I want it to run the other tests

* Specify a reason

* Install TF via conda

Was getting numpy disagreements

* Various changes for reasons I don't understand

Parsl does not like localhost, array does not like a generator

* Display environment after installation

* Specify a CPU build of tensorflow

* Try a OSX-specific environment

It'll use tensorflow-macos

* Use regular tensorflow

* Fix the numpy version

* Skip testing NFP on MacOS

TF support on Mac is just too painful

* Improve skip annotations

* Do not fork on tests

* Use RDKit from Anaconda

Trying to further reduce the number of packages from PyPI

* Use MOPAC instead of xTB in the build

* Switch to using MOPAC in examples and tests

* Only test xTB if we have it

* Temp: Fail on the first error

* Use a new version of ASE

* Do not yet test CP2K or xTB on Mac

* Use MOPAC for the example recipe

xTB not supported on Mac

* Unskip tests, allow full run

* Run the example from the command line

* Skip CPU affinity on Mac

* Use 127.0.0.1 instead of localhost

New version of Parsl only accepts IPv4 addresses

* Use MOPAC rather than xTB

* We do not install tools for all features

* Talk about the Mac installation